requirements.txt
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
docstring_parser==0.16
exceptiongroup==1.2.2
fastapi==0.88.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.84.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
idna==3.10
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.21
python-dotenv
python-dateutil==2.9.0.post0
python-multipart==0.0.20
requests==2.28.1
rsa==4.9
Shapely==1.8.5.post1
six==1.17.0
sniffio==1.3.1
starlette==0.22.0
tqdm==4.64.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.20.0